Changed from rake 11.3.0 to 10.5.0

+++ features/support/env.rb
@@ -38,5 +38,5 @@ end
  
 at_exit do
   @browser.close if @browser
-  headless.destroy
+  headless.destroy if headless
 end
